loader_test: move build-id parse logic out
So that it can be re-used in a subsequent commit.
